refactor: reorganize code in taskCard
